+++ examples/main/shared/editor/EditorPlugins.jsx
@@ -270,8 +270,8 @@ export const config = {
       metaSiteId: '538fa6c6-c953-4cdd-86c4-4b869aecf980',
       mediaRoot: 'some-mediaRoot',
     },
-    onImageEditorOpen: () => console.log('Media Studio Launched'),
     // createGalleryForMultipleImages: true,
+    onImageEditorOpen: () => console.log('Media Studio Launched'),
     // toolbar: {
     //   icons: {
     //     InsertPluginButtonIcon: MyCustomIcon,
@@ -456,14 +456,19 @@ export const config = {
     // },
     handleFileSelection: updateEntity => {
       const filenames = ['image.jpg', 'document.pdf', 'music.mp3'];
-      const name = filenames[Math.floor(Math.random() * filenames.length)];
-      const filenameParts = name.split('.');
-      const type = filenameParts[filenameParts.length - 1];
-      const data = {
-        name,
-        type,
-        url: 'http://file-examples.com/wp-content/uploads/2017/10/file-sample_150kB.pdf',
-      };
+      const multiple = false;
+      const count = multiple ? [1, 2, 3] : [1];
+      const data = [];
+      count.forEach(_ => {
+        const name = filenames[Math.floor(Math.random() * filenames.length)];
+        const filenameParts = name.split('.');
+        const type = filenameParts[filenameParts.length - 1];
+        data.push({
+          name,
+          type,
+          url: 'http://file-examples.com/wp-content/uploads/2017/10/file-sample_150kB.pdf',
+        });
+      });
       setTimeout(() => updateEntity({ data }), 500);
     },
   },